resume/requirements.txt
fastapi
python-dotenv
uvicorn
fastapi-jwt-auth
jinja2
python-multipart
google-cloud-storage
google-genai